Change "unsigned long" to "void *" to take into account LLP64 archs.

+++ devel/dmucs/files/patch-dmucs_msg.cc
@@ -5,7 +5,7 @@
  	fprintf(stderr, "Giving out %s\n", resolved_name.c_str());
  
 -	db->assignCpuToClient(cpuIpAddr, dprop_, (unsigned int) sock);
-+	db->assignCpuToClient(cpuIpAddr, dprop_, (unsigned long) sock);
++	db->assignCpuToClient(cpuIpAddr, dprop_, sock);
  #if 0
  	fprintf(stderr, "The databases are now:\n");
  	db->dump();
