+++ devel/dmucs/files/patch-main.cc
@@ -5,7 +5,7 @@
      if (Sgets(buf, BUFSIZE, sock_req) == NULL) {
  	DMUCS_DEBUG((stderr, "Socket closed: %s\n", peer2buf(sock_req, buf)));
 -	db->releaseCpu((unsigned int)sock_req);
-+	db->releaseCpu((unsigned long)sock_req);
++	db->releaseCpu(sock_req);
  	removeFd(sock_req);
  	return;
      }
